2010-12-09  Sheriff Bot  <webkit.review.bot@gmail.com>
        Unreviewed, rolling out r73616.
        http://trac.webkit.org/changeset/73616
        https://bugs.webkit.org/show_bug.cgi?id=50772

        Breaks chromium win build (Requested by hwennborg on #webkit).

        * public/WebIDBKey.h:
        (WebKit::WebIDBKey::WebIDBKey):
        * src/WebIDBKey.cpp:
        (WebKit::WebIDBKey::assign):
        (WebKit::WebIDBKey::number):
2010-12-09  Sheriff Bot  <webkit.review.bot@gmail.com>

        Unreviewed, rolling out r73616.
        http://trac.webkit.org/changeset/73616
        https://bugs.webkit.org/show_bug.cgi?id=50772

        Breaks chromium win build (Requested by hwennborg on #webkit).

        * bindings/v8/IDBBindingUtilities.cpp:
        (WebCore::createIDBKeyFromValue):
        * bindings/v8/custom/V8IDBKeyCustom.cpp:
        (WebCore::toV8):
        * storage/IDBFactoryBackendImpl.cpp:
        (WebCore::createTables):
        (WebCore::IDBFactoryBackendImpl::open):
        * storage/IDBKey.cpp:
        (WebCore::IDBKey::IDBKey):
        (WebCore::IDBKey::fromQuery):
        (WebCore::IDBKey::bind):
        (WebCore::IDBKey::bindWithNulls):
        * storage/IDBKey.h:
        (WebCore::IDBKey::create):
        (WebCore::IDBKey::number):
2010-12-09  Sheriff Bot  <webkit.review.bot@gmail.com>

        Unreviewed, rolling out r73616.
        http://trac.webkit.org/changeset/73616
        https://bugs.webkit.org/show_bug.cgi?id=50772

        Breaks chromium win build (Requested by hwennborg on #webkit).

        * storage/indexeddb/index-basics-expected.txt:
        * storage/indexeddb/index-basics.html:
        * storage/indexeddb/index-cursor.html:
        * storage/indexeddb/keyrange-expected.txt:
        * storage/indexeddb/keyrange.html:
        * storage/indexeddb/objectstore-cursor.html:


Canonical link: https://commits.webkit.org/64041@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@73633 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/storage/IDBKey.h
@@ -43,7 +43,7 @@ public:
     {
         return adoptRef(new IDBKey());
     }
-    static PassRefPtr<IDBKey> create(double number)
+    static PassRefPtr<IDBKey> create(int32_t number)
     {
         return adoptRef(new IDBKey(number));
     }
@@ -68,7 +68,7 @@ public:
         return m_string;
     }
 
-    double number() const
+    int32_t number() const
     {
         ASSERT(m_type == NumberType);
         return m_number;
@@ -88,12 +88,12 @@ public:
 
 private:
     IDBKey();
-    explicit IDBKey(double);
+    explicit IDBKey(int32_t);
     explicit IDBKey(const String&);
 
     Type m_type;
     String m_string;
-    double m_number;
+    int32_t m_number;
 };
 
 }
